refactor: moves "/URLComponent/Path/" to "/URLComponent/Path_dep/"
