+++ src/library/URLComponent/Path_dep/definition.declared.augmentation.ts
