+++ src/library/URLComponent/Path_dep/definition.declared.withAugmentation.ts
